Add recursion to FileLocator
The FileLocator uses an array of directories to search for metadata for the given class. It will then loop over this array and try to find exact matches by converting backslashes in the FQCN to dots and appending the extension.

Because of this, it will always try to do an exact match. If it cannot find that file, it will continue to the next directory.

If you need to add metadata to a lot of classes, this quickly becomes very messy. You need to create a single file for every class you want to support.

With this change, it should be easier. You can now add multiple definitions in 1 metadata file. The FileLocator will now loop every FQCN until it finds the top level namespace.

Signed-off-by: Ruud Kamphuis <ruudk@users.noreply.github.com>